feat: add changes to optional arguments and testing

+++ backend/src/routes/projects.py
@@ -3,7 +3,7 @@ from flask_restx import Resource, Namespace
 from flask_jwt_extended import jwt_required, current_user
 
 from src.extensions import db
-from src.models.group import Group
+from src.models.tutorial import Tutorial
 from src.models.project import Project, project_new_model
 from src.models.task import Task
 from src.models.user import User, UserSet
@@ -22,7 +22,12 @@ from .helpers import (
     update_db_object,
     add_db_object,
 )
-from .access_checks import check_is_member, check_project_creator
+from .access_checks import (
+    check_is_member,
+    check_project_view_access,
+    check_project_edit_access,
+    check_task_view_access,
+)
 
 # projects = Blueprint("projects", __name__)
 
@@ -78,12 +83,15 @@ class ProjectGeneral(Resource):
 
     @projects_ns.expect(project_creation_input)
     def post(self):
-        # Check user is a member of group or is the tutorial's tutor
-        group: Group = fetch_one(Group, {"id": projects_ns.payload["group_id"]})
-        check_is_member(Group, group, current_user)
+        # Check user is a member of the tutorial or is a tutor (assuming groups
+        # are now redundant)
+        tutorial: Tutorial = fetch_one(
+            Tutorial, {"id": projects_ns.payload["tutorial_id"]}
+        )
+        check_project_view_access(tutorial, current_user)
         new_project = Project(
             course_id=projects_ns.payload["course_id"],
-            group_id=projects_ns.payload["group_id"],
+            tutorial_id=projects_ns.payload["tutorial_id"],
             name=projects_ns.payload["name"],
             creator=current_user,
             subheading=projects_ns.payload.get("subheading"),
@@ -91,7 +99,7 @@ class ProjectGeneral(Resource):
             end_date=projects_ns.payload.get("end_date"),
         )
         add_db_object(Project, new_project, new_project.name)
-        new_project.add_members(members=group.members)
+        new_project.add_members(members=[current_user])
         return {}, 201
 
 
@@ -102,21 +110,23 @@ class ProjectSpecific(Resource):
     @projects_ns.marshal_with(project_fetch_one_output)
     def get(self, project_id: str):
         project: Project = fetch_one(Project, {"id": project_id})
-        check_is_member(Project, project, current_user)
+        tutorial: Tutorial = fetch_one(Tutorial, {"id": project.tutorial_id})
+        # Project can be fetched by tutorial members and tutors
+        # (can only see basic name, description of project nothing else)
+        check_project_view_access(tutorial, current_user)
         return project
 
     @projects_ns.expect(project_edit_input)
     def put(self, project_id: str):
         project: Project = fetch_one(Project, {"id": project_id})
-        # Check user is project creator or tutorial tutor
-        check_project_creator(project, current_user)
+        # Check user is project creator or tutor
+        check_project_edit_access(project, current_user, project.course_id)
         project.update(project_data=projects_ns.payload)
         return update_db_object(Project, project)
 
     def delete(self, project_id: str):
         project: Project = fetch_one(Project, {"id": project_id})
-        # Check user is project creator or tutorial tutor
-        check_project_creator(project, current_user)
+        check_project_edit_access(project, current_user, project.course_id)
         db.session.delete(project)
         db.session.commit()
         return {}, 200
@@ -129,7 +139,9 @@ class ProjectMembers(Resource):
     @projects_ns.expect(project_members_fetch_output)
     def get(self, project_id: str):
         project: Project = fetch_one(Project, {"id": project_id})
-        check_is_member(Project, project, current_user)
+        tutorial: Tutorial = fetch_one(Tutorial, {"id": project.tutorial_id})
+        # All tut members and tutors can fetch project members
+        check_project_view_access(tutorial, current_user)
         return project.members
 
     # Leave this commented out for now in case a route is needed, but joining a
@@ -149,6 +161,7 @@ class ProjectTasks(Resource):
     @projects_ns.marshal_list_with(task_fetch_all_output)
     def get(self, project_id: str):
         project: Project = fetch_one(Project, {"id": project_id})
-        check_is_member(Project, project, current_user)
+        # Must be project member or tutor to view project tasks
+        check_task_view_access(project, current_user)
         tasks: list[Task] = fetch_all(Task)
         return [tsk for tsk in tasks if tsk.project_id == project_id]
